fix: update ISERR macro to ensure proper type casting for error checking

+++ src/kernel.h
@@ -15,4 +15,4 @@ struct paging_desc *kernel_desc();
 
 #define ERROR(value) (void *)((intptr_t)(value))
 #define ERROR_I(value) (int)((intptr_t)(value))
-#define ISERR(value) ((int)value < 0)
+#define ISERR(value) ((int)(((int64_t)value) < 0))
